platina-mk1.bin: Relocate to 0xffedf680 make room for larger bootblock
Enabling CONFIG_USE_OPTION_TABLE expands the bootblock from 1264
bytes to 2184 bytes. Move the FSP down in ROM to make room for this
expension.

Created using BCT as the following:

~/co/BCT/bct/bct -i BROADWELLDE_FSP.bin -b platina-mk1.bsf -o platina-mk1.temp
~/co/BCT/bct/bct -i platina-mk1.temp -r 0xffedf680 -o platina-mk1.bin

Note the two steps. In my testing, if you attempt to relocate and change
settings in one step, it will not change the settings.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>